bookshelf/commons/controller_test.go
package commons

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/labstack/echo/v4"
	"github.com/stretchr/testify/assert"
)

func TestHealth(t *testing.T) {
	e := echo.New()
	req := httptest.NewRequest(echo.GET, "/api/v1/health", nil)
	rec := httptest.NewRecorder()
	c := e.NewContext(req, rec)

	if assert.NoError(t, Health(c)) {
		assert.Equal(t, http.StatusOK, rec.Code)
		assert.Equal(t, rec.Body.String(), "OK")
	}
}

func TestSampleToken(t *testing.T) {
	e := echo.New()
	req := httptest.NewRequest(echo.GET, "/sample-token", nil)
	rec := httptest.NewRecorder()
	c := e.NewContext(req, rec)

	if assert.NoError(t, GenerateSampleToken(c)) {
		assert.Equal(t, http.StatusOK, rec.Code)
	}
}
